+++ src/low/msp-asm/relic_fb_mul_low.c
@@ -150,7 +150,7 @@ void fb_muln_low(dig_t *c, dig_t *a, dig_t *b) {
 	}
 }
 */
-void fb_muld_low(dig_t *c, const dig_t *a, const dig_t *b, int size) {
+void fb_muld_low(dig_t *c, const dig_t *a, const dig_t *b, size_t size) {
 	dv_t table[16];
 	dig_t u, *tmpc, r0, r1, r2, r4, r8;
 	const dig_t *tmpa;
