+++ src/low/x64-asm-6l/relic_bn_mul_low.c
@@ -40,19 +40,19 @@
 /* Public definitions                                                         */
 /*============================================================================*/
 
-dig_t bn_mula_low(dig_t *c, const dig_t *a, dig_t digit, int size) {
+dig_t bn_mula_low(dig_t *c, const dig_t *a, dig_t digit, size_t size) {
 	return mpn_addmul_1(c, a, size, digit);
 }
 
-dig_t bn_mul1_low(dig_t *c, const dig_t *a, dig_t digit, int size) {
+dig_t bn_mul1_low(dig_t *c, const dig_t *a, dig_t digit, size_t size) {
 	return mpn_mul_1(c, a, size, digit);
 }
 
-void bn_muln_low(dig_t *c, const dig_t *a, const dig_t *b, int size) {
+void bn_muln_low(dig_t *c, const dig_t *a, const dig_t *b, size_t size) {
 	mpn_mul_n(c, a, b, size);
 }
 
-void bn_muld_low(dig_t *c, const dig_t *a, int sizea, const dig_t *b, int sizeb,
+void bn_muld_low(dig_t *c, const dig_t *a, size_t sizea, const dig_t *b, size_t sizeb,
 		int low, int high) {
 	(void)low;
 	(void)high;
